requirements.txt
numpy
pyaudio
soundfile
librosa
openai
faster-whisper
requests
ffmpeg
tiktoken
pyyaml
uvicorn
websockets
fastapi
pydantic_settings
starlette
sse_starlette
starlette_context
whisper_timestamped
elevenlabs
vllm
pynvml
bitsandbytes